Add car review dataset and reference translation
This commit adds the `car_reviews.csv` and `reference_translation.txt` files to the `data` folder. The `car_reviews.csv` file contains car reviews and their class (positive or negative), and the `reference_translation.txt` file contains the Spanish versions of some reviews. These files are used in the Analyzing-Car-Reviews-with-LLMs notebook to build and evaluate the chatbot model for various NLP tasks.